fix gcc-7 / cuda10 compilation

+++ cuda/lltm_cuda_kernel.cu
@@ -25,7 +25,7 @@ __device__ __forceinline__ scalar_t d_tanh(scalar_t z) {
 
 template <typename scalar_t>
 __device__ __forceinline__ scalar_t elu(scalar_t z, scalar_t alpha = 1.0) {
-  return fmax(0.0, z) + fmin(0.0, alpha * (exp(z) - 1.0));
+  return fmaxf(0.0, z) + fminf(0.0, alpha * (exp(z) - 1.0));
 }
 
 template <typename scalar_t>
